refactor: extract utils from different layers
