+++ app/src/main/java/io/github/chhabra_dhiraj/spaceflightnews/util/datetime/LocalDateTimeParser.kt
@@ -1,4 +1,4 @@
-package io.github.chhabra_dhiraj.spaceflightnews.feature.article.data.util
+package io.github.chhabra_dhiraj.spaceflightnews.util.datetime
 
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
