+++ app/src/main/java/io/github/chhabra_dhiraj/spaceflightnews/util/error/ErrorTextUtil.kt
@@ -1,8 +1,8 @@
-package io.github.chhabra_dhiraj.spaceflightnews.feature.article.presentation.util
+package io.github.chhabra_dhiraj.spaceflightnews.util.error
 
 import androidx.annotation.StringRes
 import io.github.chhabra_dhiraj.spaceflightnews.R
-import io.github.chhabra_dhiraj.spaceflightnews.feature.article.domain.util.DataError
+import io.github.chhabra_dhiraj.spaceflightnews.util.network.DataError
 
 @StringRes
 fun DataError.getErrorRes(): Int {
