+++ app/src/main/java/io/github/chhabra_dhiraj/spaceflightnews/util/network/Error.kt
@@ -1,0 +1,3 @@
+package io.github.chhabra_dhiraj.spaceflightnews.util.network
+
+sealed interface Error
